Lo que hice ayer antes de irme a dormir (ya ni me acuerdo)

+++ src/main/java/bryan/travelAPP/repository/AirportRepository.java
@@ -3,5 +3,5 @@ package bryan.travelAPP.repository;
 import bryan.travelAPP.domain.Airport;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AirportRepository extends CrudRepository<Airport, Long> {
+public interface AirportRepository extends CrudRepository<Airport, Integer> {
 }
