+++ src/main/java/bryan/travelAPP/repository/FlightRepository.java
@@ -3,5 +3,5 @@ package bryan.travelAPP.repository;
 import bryan.travelAPP.domain.Flight;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FlightRepository extends CrudRepository<Flight, Long> {
+public interface FlightRepository extends CrudRepository<Flight, Integer> {
 }
