+++ src/main/java/bryan/travelAPP/repository/PassengerRepository.java
@@ -3,5 +3,5 @@ package bryan.travelAPP.repository;
 import bryan.travelAPP.domain.Passenger;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PassengerRepository extends CrudRepository<Passenger, Long> {
+public interface PassengerRepository extends CrudRepository<Passenger, Integer> {
 }
