+++ src/main/java/bryan/travelAPP/repository/SeatRepository.java
@@ -3,5 +3,5 @@ package bryan.travelAPP.repository;
 import bryan.travelAPP.domain.Seat;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SeatRepository extends CrudRepository<Seat, Long> {
+public interface SeatRepository extends CrudRepository<Seat, Integer> {
 }
